ref: Save progress componentizing
