+++ client/src/components/atoms/SubmissionCommentSection.vue
@@ -1,0 +1,266 @@
+<template>
+  <section class="comments">
+    <div class="comments-wrapper">
+      <h3 class="text-h1">Overall Comments</h3>
+      <q-card>
+        <q-card-section>
+          <div class="row items-start justify-between content-stretch no-wrap">
+            <div class="col-grow content-between">
+              <div class="text-h4">Vitae Congue</div>
+              <div>
+                <small>February 17th, 2021 at 6:35pm</small>
+              </div>
+            </div>
+
+            <div class="col-auto">
+              <q-btn
+                color="grey-7"
+                round
+                flat
+                icon="more_vert"
+                aria-label="Comment Menu"
+              >
+                <q-menu cover auto-close>
+                  <q-list>
+                    <q-item clickable>
+                      <q-item-section>Remove Card</q-item-section>
+                    </q-item>
+                    <q-item clickable>
+                      <q-item-section>Send Feedback</q-item-section>
+                    </q-item>
+                    <q-item clickable>
+                      <q-item-section>Share</q-item-section>
+                    </q-item>
+                  </q-list>
+                </q-menu>
+              </q-btn>
+            </div>
+          </div>
+        </q-card-section>
+        <q-separator />
+        <q-card-section class="q-pb-none">
+          <p>
+            Vitae semper quis lectus nulla at volutpat. Eleifend quam adipiscing
+            vitae proin sagittis. Tellus molestie nunc non blandit massa. Odio
+            tempor orci dapibus ultrices in. Condimentum id venenatis a
+            condimentum vitae sapien pellentesque habitant. Auctor augue mauris
+            augue neque gravida in. Etiam sit amet nisl purus in. Fringilla ut
+            morbi tincidunt augue. Morbi tincidunt ornare massa eget egestas
+            purus viverra accumsan. Sed odio morbi quis commodo. Velit euismod
+            in pellentesque massa. Massa massa ultricies mi quis hendrerit dolor
+            magna eget.
+          </p>
+          <p>
+            Sed sed risus pretium quam vulputate. Amet mauris commodo quis
+            imperdiet massa tincidunt nunc. Consequat mauris nunc congue nisi
+            vitae suscipit tellus mauris a. Non nisi est sit amet facilisis
+            magna. Turpis massa tincidunt dui ut ornare lectus. Vel risus
+            commodo viverra maecenas accumsan lacus vel facilisis. Mauris cursus
+            mattis molestie a iaculis at. Aenean sed adipiscing diam donec. Nisl
+            nunc mi ipsum faucibus. Sed elementum tempus egestas sed sed risus.
+            Risus pretium quam vulputate dignissim suspendisse in. Ut faucibus
+            pulvinar elementum integer. Volutpat ac tincidunt vitae semper quis.
+            Cras semper auctor neque vitae tempus. Malesuada bibendum arcu vitae
+            elementum curabitur vitae nunc sed. Tortor vitae purus faucibus
+            ornare suspendisse sed. Turpis tincidunt id aliquet risus feugiat.
+            Mauris augue neque gravida in fermentum et sollicitudin ac orci. In
+            pellentesque massa placerat duis ultricies lacus sed.
+          </p>
+        </q-card-section>
+        <q-card-actions class="q-pa-md">
+          <q-btn bordered color="primary">Reply</q-btn>
+        </q-card-actions>
+      </q-card>
+      <div class="q-mx-md q-mb-md">
+        <q-separator />
+        <q-card flat square class="bg-grey-1">
+          <q-card-section>
+            <div class="row items-start no-wrap">
+              <div class="col">
+                <div class="text-subtitle1">Egestas</div>
+                <div>
+                  <small>
+                    <q-icon size="sm" name="subdirectory_arrow_right" />
+                    <span>Reply to Magna Fringilla</span>
+                  </small>
+                </div>
+                <small> February 18th, 2021 at 6:35pm</small>
+              </div>
+
+              <div class="col-auto">
+                <q-btn
+                  color="grey-7"
+                  round
+                  flat
+                  icon="more_vert"
+                  aria-label="Reply Comment Menu"
+                >
+                  <q-menu cover auto-close>
+                    <q-list>
+                      <q-item clickable>
+                        <q-item-section>Remove Card</q-item-section>
+                      </q-item>
+                      <q-item clickable>
+                        <q-item-section>Send Feedback</q-item-section>
+                      </q-item>
+                      <q-item clickable>
+                        <q-item-section>Share</q-item-section>
+                      </q-item>
+                    </q-list>
+                  </q-menu>
+                </q-btn>
+              </div>
+            </div>
+          </q-card-section>
+
+          <q-card-section class="q-py-none">
+            <p>
+              Sagittis eu volutpat odio facilisis. Vitae congue eu consequat ac.
+              Cursus sit amet dictum sit amet. Nibh tellus molestie nunc non
+              blandit massa enim. Et tortor consequat id porta nibh venenatis.
+              Dictum at tempor commodo ullamcorper. Placerat orci nulla
+              pellentesque dignissim. Rhoncus dolor purus non enim praesent
+              elementum facilisis.
+            </p>
+          </q-card-section>
+
+          <q-card-actions class="q-pa-md">
+            <q-btn bordered color="primary">Reply</q-btn>
+          </q-card-actions>
+
+          <q-separator />
+        </q-card>
+        <q-card flat square class="bg-grey-1">
+          <q-card-section>
+            <div class="row items-start no-wrap">
+              <div class="col">
+                <div class="text-subtitle1">Nibh Mauris</div>
+                <div>
+                  <small>
+                    <q-icon size="sm" name="subdirectory_arrow_right" />
+                    <span>Reply to Egestas</span>
+                  </small>
+                </div>
+                <small> February 18th, 2021 at 6:35pm</small>
+              </div>
+
+              <div class="col-auto">
+                <q-btn
+                  color="grey-7"
+                  round
+                  flat
+                  icon="more_vert"
+                  aria-label="Reply Comment Menu"
+                >
+                  <q-menu cover auto-close>
+                    <q-list>
+                      <q-item clickable>
+                        <q-item-section>Remove Card</q-item-section>
+                      </q-item>
+                      <q-item clickable>
+                        <q-item-section>Send Feedback</q-item-section>
+                      </q-item>
+                      <q-item clickable>
+                        <q-item-section>Share</q-item-section>
+                      </q-item>
+                    </q-list>
+                  </q-menu>
+                </q-btn>
+              </div>
+            </div>
+          </q-card-section>
+
+          <q-card-section class="q-py-none">
+            <p>Dictum at tempor commodo.</p>
+          </q-card-section>
+
+          <q-card-actions class="q-pa-md">
+            <q-btn bordered color="primary">Reply</q-btn>
+          </q-card-actions>
+
+          <q-separator />
+        </q-card>
+      </div>
+
+      <q-card class="q-mb-md">
+        <q-card-section>
+          <div class="row items-start justify-between content-stretch no-wrap">
+            <div class="col-grow content-between">
+              <div class="text-h4">Amet Nisl Purus</div>
+              <div>
+                <small>February 17th, 2021 at 6:35pm</small>
+              </div>
+            </div>
+
+            <div class="col-auto">
+              <q-btn
+                color="grey-7"
+                round
+                flat
+                icon="more_vert"
+                aria-label="Comment Menu"
+              >
+                <q-menu cover auto-close>
+                  <q-list>
+                    <q-item clickable>
+                      <q-item-section>Remove Card</q-item-section>
+                    </q-item>
+                    <q-item clickable>
+                      <q-item-section>Send Feedback</q-item-section>
+                    </q-item>
+                    <q-item clickable>
+                      <q-item-section>Share</q-item-section>
+                    </q-item>
+                  </q-list>
+                </q-menu>
+              </q-btn>
+            </div>
+          </div>
+        </q-card-section>
+        <q-separator />
+        <q-card-section class="q-pb-none">
+          <p>
+            Vitae semper quis lectus nulla at volutpat. Eleifend quam adipiscing
+            vitae proin sagittis. Tellus molestie nunc non blandit massa. Odio
+            tempor orci dapibus ultrices in. Condimentum id venenatis a
+            condimentum vitae sapien pellentesque habitant. Auctor augue mauris
+            augue neque gravida in. Etiam sit amet nisl purus in. Fringilla ut
+            morbi tincidunt augue. Morbi tincidunt ornare massa eget egestas
+            purus viverra accumsan. Sed odio morbi quis commodo. Velit euismod
+            in pellentesque massa. Massa massa ultricies mi quis hendrerit dolor
+            magna eget.
+          </p>
+          <p>
+            Sed sed risus pretium quam vulputate. Amet mauris commodo quis
+            imperdiet massa tincidunt nunc. Consequat mauris nunc congue nisi
+            vitae suscipit tellus mauris a. Non nisi est sit amet facilisis
+            magna. Turpis massa tincidunt dui ut ornare lectus. Vel risus
+            commodo viverra maecenas accumsan lacus vel facilisis. Mauris cursus
+            mattis molestie a iaculis at. Aenean sed adipiscing diam donec. Nisl
+            nunc mi ipsum faucibus. Sed elementum tempus egestas sed sed risus.
+            Risus pretium quam vulputate dignissim suspendisse in. Ut faucibus
+            pulvinar elementum integer. Volutpat ac tincidunt vitae semper quis.
+            Cras semper auctor neque vitae tempus. Malesuada bibendum arcu vitae
+            elementum curabitur vitae nunc sed. Tortor vitae purus faucibus
+            ornare suspendisse sed. Turpis tincidunt id aliquet risus feugiat.
+            Mauris augue neque gravida in fermentum et sollicitudin ac orci. In
+            pellentesque massa placerat duis ultricies lacus sed.
+          </p>
+        </q-card-section>
+        <q-card-actions class="q-pa-md">
+          <q-btn bordered color="primary">Reply</q-btn>
+        </q-card-actions>
+      </q-card>
+    </div>
+  </section>
+</template>
+
+<style lang="sass" scoped>
+.comments
+  background-color: #efefef
+
+.comments-wrapper
+  max-width: 700px
+  margin: 0 auto
+  padding: 10px 60px 60px
+</style>
